==> 02 - cleaning data in df.ipynb <==
df = pd.read_csv('rowDF.csv')

#creating clean by word function:
def clean_by_word(df_column , fromWord ,toWord , addons_list):
    new_list = list ()
    clean_new_list = list()
    for index, w in enumerate(df[df_column]):
        df[df_column][index+1].strip()
        word =  df[df_column][index+1]
        result = word.find(fromWord)
        word2 =  df[df_column][index+1][result:]
        new_list.append(word2)
        result2 = word2.find(toWord)
        clean_new_list.append(new_list[index][:result2])
        df[df_column][index+1] = clean_new_list[index]

    for addon in addons_list:
        df[df_column] = df[df_column].str.replace(addon, '')

# change dtype object >> str:
df['GameTitles'] = df['GameTitles'].astype(str)
df['ReleaseYear'] = df['ReleaseYear'].astype(str)
df['Price'] = df['Price'].astype(str)
df['SteamPositivePrecent'] = df['SteamPositivePrecent'].astype(str)
df['SteamUserReviews'] = df['SteamUserReviews'].astype(str)
df['Metascore'] = df['Metascore'].astype(str)


#remove useless chars:
spec_chars = ["!",'"',"#","%","&","(",")",
              "*","+",",","-","/",";","<",
              "=",">","?","@","[","\\","]","^",
              "`","{","|","}","~" ,"™" ,"®"]
rev_price_addons = ["span" , "class" , "title" , "div" , "strike" , "col" 
                ,"search_price" , "search_released", "responsive_secondrow",
                "search_reviewscore" , "search_review_summary" , "positive" ,
                "data", "tooltip" , "html" , "\\n" , "Positive" , "Very",
                "lt" , "br" ,"gt" , "Overwhelmingly" , "Mostly" , "mixed",
                "Mixed" , "of" , "discounted" , "style" , "or" , "888888" ,"user reviews f this game are",
                "This product has experienced" , ". one  me periods  ftopic review activity.",
                 "Based on your ", "preferences the review" , "s within these periods" ,
                 "    have been excluded" , "  from this products " , "Review Sce" , '< ="platfm_img win"></>'] 

html_addons = ["span" , "class" , "title" , "div" , "strike" , "col" 
                ,"search_price" , "search_released", "responsive_secondrow","omig",
                "search_reviewscore" , "search_review_summary" ,
                "data" , "\t\t\t\t\t\t\t\t\t\n\n\nmetacritic\nRead" , "Critic Reviews","NA\n\nmetacritic\nNot yet reviewed.","Soon","e 16","ment","ouce"] 

more_addons = ["Publisher:" , "\\n" , "Genre:" , "Franchise:" , "Storage:" ,", Free to Play" , "Free to Play," ,", Early Access",'Atleast']

pos_rev_addons = ["\\n" ,'<div class="col search_reviewscore responsive' ,'_secondrow"><span class="search_review_summary',
                    ' positive" data-tooltip-html="Very Positive&lt;br&gt;' , ' user reviews for this game are positive."></span></div>',
                    'positive" data-tooltip-html="Overwhelmingly','Positive&lt;br&gt;','mixed" data-tooltip-html="Mixed&lt;br&gt;', 'Storage',':',' ','fthe',".",
                    'positive" data-tooltip-html="Mostly ', ' user reviews for this game ','are positive.&lt;br&gt;&lt;br&gt;This product has experienced one or more periods of off-topic review activity.',
                    '  Based on your preferences, the reviews within these periods have been excluded from this product\'s Review Score."></span></div>','"',
                    ' user reviews for this software are positive.','></span></div>' ,'of the ' ,',' ,'_secondrow"></div>', "_secondrow></div>"," or " ,"+" ,"m"," user reviews for gaes in this bundle are positive.", "></span></div>",
                    "userreviewsforthisgaearepositive",".&lt;br&gt;&lt;br&gt;Thisproducthasexperiencedoneororeperiodsofoff-topicreviewactivity.Basedonyourpreferencesthereviewswithintheseperiodshavebeenexcludedfrothisp",
                    "roduct'sReviewScore.","&lt;br&gt;&lt;br&gt;Thisproducthasexperiencedoneororeperiodsofoff-topicreviewactivityBasedonyourpreferencesthereviewswithintheseperiodshavebeenexcludedfrothispro",
                    "duct'sReviewScore" ,"userreviewsforthissoftwarearepositive" ,'userreviewsforgaesinthisbundlearepositive','Atleast',
                    'Mindestens','userreviewsforthisvideoarepositive']      

for char in spec_chars:
    df['GameTitles'] = df['GameTitles'].str.replace(char, '')
    df['ReleaseYear'] = df['ReleaseYear'].str.replace(char, '')
    df['Price'] = df['Price'].str.replace(char, '')
    df['SteamPositivePrecent'] = df['SteamPositivePrecent'].str.replace(char, '')    

for addon in html_addons:
    df['GameTitles'] = df['GameTitles'].str.replace(addon, '')
    df['ReleaseYear'] = df['ReleaseYear'].str.replace(addon, '')
    df['Metascore'] = df['Metascore'].str.replace(addon, '')

for addon in rev_price_addons:
    df['Price'] = df['Price'].str.replace(addon, '')
    df['SteamPositivePrecent'] = df['SteamPositivePrecent'].str.replace(addon, '')

for posrev in pos_rev_addons:
    df['SteamUserReviews'] = df['SteamUserReviews'].str.replace(posrev, '')

#striping and cleaning columns:
for i in range(1,len(game_title)+1):
    df['GameTitles'][i] = df['GameTitles'][i].strip()
    df['ReleaseYear'][i] = df['ReleaseYear'][i].strip()
    df['Price'][i] = df['Price'][i].strip()
    df['SteamPositivePrecent'][i] = df['SteamPositivePrecent'][i].strip()


#creating nums list of int >> str
rate_list = list()
for i in range (0,101):
    rate_list.append(i)   

str_rate_list = [str(i) for i in rate_list]

#cleaning reviews:
i=1
for x in df['SteamPositivePrecent']:
    df['SteamPositivePrecent'][i]=x[:4].strip()
    if df['SteamPositivePrecent'][i] not in str_rate_list:
        df['SteamPositivePrecent'][i]='-1'
    i+=1
df['SteamPositivePrecent'] = df['SteamPositivePrecent'].astype(int)

#cleaning from details by word:
clean_by_word('Publisher' , 'Publisher:' , '\n\n', more_addons)
clean_by_word('Genre' , 'Genre:' ,'\n\n', more_addons)
clean_by_word('Franchise' , 'Franchise:' ,'\n\n', more_addons)       

#cleaning Storage:

#removing GB/MB:

new_list = list ()
clean_new_list = list()

for index, w in enumerate(df['Storage(GB)']):
    df['Storage(GB)'][index+1].strip()
    word =  df['Storage(GB)'][index+1]
    result = word.find('Storage:')
    word2 =  df['Storage(GB)'][index+1][result+8:]
    new_list.append(word2)
    result2 = word2.find('B')
    clean_new_list.append(new_list[index][:result2+1])
    if (result >= 0):
        df['Storage(GB)'][index+1] = clean_new_list[index]
    if (word2 == 0 or result == -1):
        df['Storage(GB)'][index+1] = 0

Storage_addons = ['At least','Mindestens', '+','-','/',']','NaN','nan',
'[',':','*',' ','.384\t\t\t\t\t1.384']

for addon in Storage_addons:
    df['Storage(GB)'] = df['Storage(GB)'].str.replace(addon, '')


df['Storage(GB)'] = df['Storage(GB)'].astype(str)

#------------------------------------

# converting Storage(GB) to numbers:

for index, s in enumerate(df['Storage(GB)']):
    df['Storage(GB)'][index+1].strip()
    word =  df['Storage(GB)'][index+1]
    result = word.find('MB')
    if result >= 0:
        df['Storage(GB)'][index+1]=df['Storage(GB)'][index+1][:-2].strip()
    else:
        df['Storage(GB)'][index+1]=df['Storage(GB)'][index+1][:-2]
        df['Storage(GB)'][index+1]=df['Storage(GB)'][index+1]+"000"

    if (df['Storage(GB)'][index+1] == 'n000'):
        df['Storage(GB)'][index+1] = 0

df['Storage(GB)'] = df['Storage(GB)'].astype(float)

for index, s in enumerate(df['Storage(GB)']):
    df['Storage(GB)'][index+1]=df['Storage(GB)'][index+1]/1000

#cleaning ReleaseYear:

for index,x in enumerate(df['ReleaseYear']):
    counter = len(x)
    df['ReleaseYear'][index+1] = df['ReleaseYear'][index+1][counter-4:].strip()
    if df['ReleaseYear'][index+1] == '':
        df['ReleaseYear'][index+1] = 0
    try:
        n = int(df['ReleaseYear'][index+1])
    except:
        df['ReleaseYear'][index+1] = 0

df['ReleaseYear'] = df['ReleaseYear'].astype(int)

#cleaning Genre:
new_list = list ()
for index, w in enumerate(df['Genre']):
    df['Genre'][index+1].strip()
    word =  df['Genre'][index+1]
    result = word.find(',')
    if result == -1:
         df['Genre'][index+1] = word
         continue
    if result >= 0:
        word2 =  df['Genre'][index+1][:result]
        df['Genre'][index+1] = word2       

#cleaning Price:
new_list = list ()
for index, w in enumerate(df['Price']):
    df['Price'][index+1].strip()
    word =  df['Price'][index+1]
    result = word.find('₪')
    result3 = word.find('Free')
    if word == '':
        new_list.append('0')
        continue
    if result3 > -1:
         new_list.append('0')
         continue
    if result >= 0:
        word2 =  df['Price'][index+1][result:]
        result2 = word2[1:].find('₪')
        if result2 > -1:
            new_list.append(word2[1:result2+1])
        else:
            new_list.append(word2[1:])

for new_price in new_list:
    df['Price'] = new_list

df['Price'] = df['Price'].astype(float)

#cleaning PositiveReviews:
for index, r in enumerate(df['SteamUserReviews']):
    df['SteamUserReviews'][index+1].strip()
    word =  df['SteamUserReviews'][index+1]
    result = word.find('%')
    if result >= 0:
        df['SteamUserReviews'][index+1]=df['SteamUserReviews'][index+1][result+2:].strip()
    if r == '':
        df['SteamUserReviews'][index+1] = '0'
df['SteamUserReviews'] = df['SteamUserReviews'].astype(int)

# df.to_csv('cleanDF.csv')
#df = pd.read_csv('cleanDF.csv')
