==> 04 - visualisation and EDA.ipynb <==
cor = df_copy.corr(method ='pearson')
ax = sns.heatmap(cor,color='w')

# plt functions:

def one_dim_plot(sr, plot_type, axis):
    sr.plot(kind=plot_type , ax=axis)

def get_frequent_elements(df, col_name, num_top_elements):
    topele = df[col_name].value_counts()[:num_top_elements]
    topele = topele.sort_index()
    return topele

def plot_frequent_elements(df, df_in_params):
    fig, axes = plt.subplots(1,df_in_params.shape[0], figsize=(20,5))
    for i in range(df_params.shape[0]):
        x=get_frequent_elements(df,df_params['col_name'][i],df_params['num_top_elements'][i])
        one_dim_plot(x,df_params['plot_type'][i],axes[i])

def cross_tabulation(df, col_name, other_col_name):
    ct1 = pd.crosstab(df[col_name],df[other_col_name],normalize='index')
    return ct1

def plot_cross_tabulation(df, col_names, other_col_name):
    i=0
    fig, axes = plt.subplots(1,2, figsize=(20,5))
    for val in col_names:
        new_df= cross_tabulation(df,val,other_col_name)
        new_df.plot(ax=axes[i])
        i=i+1

#---------------------------


params = {"ytick.color" : "w",
          "xtick.color" : "w",
          "axes.labelcolor" : "w",
          "axes.edgecolor" : "w"}

plt.rcParams.update(params)

# bar and line plots:
df_params = pd.DataFrame({'plot_type': ['bar', 'line'], 
                          'col_name': ['ReleaseYear', 'Storage(GB)'],'num_top_elements': [30,50]})
plot_frequent_elements(df_copy, df_params)

#colorful dots plot:
fig = plt.figure(figsize=(12,7))
sns.scatterplot(x='Price' , y='Storage(GB)' , size='metaUsersScore' , hue='Genre',sizes=(2,100) ,data=df_copy)

fig = plt.figure(figsize=(12,7))
sns.scatterplot(x='metaUsersScore' , y='SteamPositivePrecent' , size='Price' , hue='Genre',sizes=(2,100) ,data=df_copy)

fig = plt.figure(figsize=(12,7))
sns.scatterplot(x='Metascore' , y='SteamPositivePrecent' , size='Price' , hue='Genre',sizes=(2,100) ,data=df_copy)

#nice new scatter plt
fig, axs = plt.subplots(ncols=3, figsize=(30,8))
sns.regplot(x="SteamPositivePrecent", y="metaUsersScore", data=df_copy, ax=axs[0], color='r')
sns.regplot(x="Metascore", y="metaUsersScore", data=df_copy, ax=axs[1], color='b')
sns.regplot(x="Price", y="Storage(GB)", data=df_copy, ax=axs[2], color='g')
axs[0].set_title('Steam Positive Precent and metascores Users Score (Top platforms)', fontsize=10, color='w')
axs[1].set_title('Critic Metascore and Metascores Users Score', fontsize=10, color='w')
axs[2].set_title('Price and Storage in (GB)', fontsize=10, color='w')
plt.show()

#pariplot:
sns.pairplot(df_copy[['metaUsersScore','ReleaseYear','Metascore','SteamUserReviews','SteamPositivePrecent','Storage(GB)','Price']])
