==> 05 - linear regresion and prediction.ipynb <==
final_df = df_copy.copy(deep=True)
final_df = final_df.drop(columns=['GameTitles','Publisher','Developer','Franchise'])
le=LabelEncoder()
final_df['Genre'] = le.fit_transform(final_df['Genre'])
#----------------------------------
#normalizinig the new df:
normalized_df=(final_df-final_df.mean())/final_df.std()

#fixing the final score:
#final_df['metaUsersScore_Metascore_ratio']=final_df['numOfLabels']/final_df['SteamUserReviews']

#creating the prediction:
x_df = final_df.drop(columns=['metaUsersScore'])
y_df = final_df['metaUsersScore']

l = LinearRegression()
x_train , x_test , y_train , y_test = train_test_split(x_df,y_df, test_size=0.2 , random_state = 8435)
l.fit(x_train,y_train)
pred = l.predict(x_test)

print('Coefficients:',l.coef_)
print('Intercept:',l.intercept_)
print('Mean squered error (MSE): %.2f' % mean_squared_error(y_test , pred))
print('Coefficient of determination (R^2): %.2f' % r2_score(y_test, pred),'/ 1.0')
print('True Score:',l.score(x_test,y_test))
#final plot:
sns.scatterplot(y_test,pred,alpha=0.6)

l.score(x_test,y_test)
